updated user model and controller
